+++ src/features/products/components/all-products/product-action-bar/productActionBar.types.ts
@@ -1,0 +1,3 @@
+import type { ProductMenuProps } from '@/features/products/components/all-products/product-menu/productMenu.types';
+
+export type ProductActionBarProps = ProductMenuProps;
