--- src/features/products/components/product-action-bar/productActionBar.types.ts
@@ -1,3 +1,0 @@
-import type { ProductMenuProps } from '@/features/products/components/product-menu/productMenu.types';
-
-export type ProductActionBarProps = ProductMenuProps;
